pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "python-binance>=1.0.36",
    "questionary>=2.1.1"
]
